+++ UnitTests/BraintreeThreeDSecureTests/BTPaymentFlowClient+ThreeDSecure_Tests.swift
@@ -22,7 +22,6 @@ class BTPaymentFlowClient_ThreeDSecure_Tests: XCTestCase {
 
         threeDSecureRequest.nonce = "fake-card-nonce"
         threeDSecureRequest.amount = 9.97
-        threeDSecureRequest.versionRequested = .version2
         threeDSecureRequest.dfReferenceID = "df-reference-id"
         threeDSecureRequest.accountType = .credit
         threeDSecureRequest.challengeRequested = true